# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/CACA/AttributeTypes": {
            "get": {
                "tags": [
                    "AttrTypeCAA"
                ],
                "summary": "Show all rows of attributes_types_caa table",
                "description": "display all rows of attributes_types_caa table",
                "operationId": "42af6cc87b9582f068ed930bf389e159",
                "responses": {
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/CACA/AttributeType": {
            "put": {
                "tags": [
                    "AttrTypeCAA"
                ],
                "summary": "update the AttributeType's id and CACA's id",
                "description": "update the AttributeType's id and CACA's id by admin",
                "operationId": "0a2ba9a6ae5d50305adc376aa6ecdc62",
                "requestBody": {
                    "description": "Pass AttributeType id and CACA id",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "attribute_type_id",
                                    "caa_id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "format": "id",
                                        "example": "12"
                                    },
                                    "attribute_type_id": {
                                        "type": "integer",
                                        "format": "name",
                                        "example": "1"
                                    },
                                    "caa_id": {
                                        "type": "integer",
                                        "format": "name",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, your data not supported."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not_Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, your data not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "AttrTypeCAA"
                ],
                "summary": "insert the AttributeType's id and CACA's id",
                "description": "insert the ID of the AttributeType and CACA",
                "operationId": "81f2af4992dcebc3b174268864bfb1a0",
                "requestBody": {
                    "description": "Pass AttributeType id and CACA id",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "attribute_type_id",
                                    "caa_id"
                                ],
                                "properties": {
                                    "attribute_type_id": {
                                        "type": "integer",
                                        "format": "id",
                                        "example": "1"
                                    },
                                    "caa_id": {
                                        "type": "integer",
                                        "format": "id",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "object_created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "AttrTypeCAA"
                ],
                "summary": "delete the row of the attributes_types_caa table that is determined",
                "description": "delete row of the attributes_types_caa table by admin",
                "operationId": "f49ef98479a0127c629c75c713ec9779",
                "requestBody": {
                    "description": "delete from attributes_types_caa table",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "format": "id",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Not_Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, your data not supported."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}